+++ changelog.md
@@ -5,11 +5,15 @@
 - `DiscordHTTPClient.guild_get` now supports query string parameters.
 - Add `Client.guild_get` even tho `Client.guild_sync` exists, just to request guild specific data
     (especially user counts).
+- Add `Permission.start_embedded_activities`.
+
+##### ext.slash
+
+- `InteractionResponse` will not retrieve message if used within a `return` expression (Saving up 1 request usually).
 
 #### Bug Fixes
 
 - `Message._late_init` had a few bad checks, making embed and such fields to not update if needed.
-
 
 ## 1.1.102 *\[2021-09-??\]*